DTOs

+++ pruebaAgencia/src/main/java/ar/edu/utn/frc/pruebas/dto/PosicionVehiculoDTO.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class PosicionVehiculoDTO {
     private int vehiculoId;
-    private int latitud;
-    private int longitud;
+    private double latitud;
+    private double longitud;
 }
